Reorganizes CardContents and EdgenoteCard as their own modules
