+++ app/components/card/CitationEntity.jsx
