+++ app/components/card/CitationTooltip.jsx
