+++ app/components/card/EdgenoteEntity.jsx
