+++ app/components/card/LinkEntity.jsx
